+++ src/OpenLoco/CMakeLists.txt
@@ -388,8 +388,9 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Version.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Viewport.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/ViewportManager.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi/CompanyAi.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi/CompanyAiPlaceVehicle.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/Company.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyManager.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/Industry.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/IndustryManager.cpp"
@@ -712,8 +713,9 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/VehicleManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Viewport.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/ViewportManager.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi/CompanyAi.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi/CompanyAiPlaceVehicle.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/Company.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyAi.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/CompanyRecords.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/World/Industry.h"
