+++ src/OpenLoco/src/World/CompanyAi/CompanyAi.h
