+++ src/OpenLoco/src/World/CompanyAi/CompanyAiPlaceVehicle.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace OpenLoco
+{
+    struct Company;
+}
+
+namespace OpenLoco::CompanyAi
+{
+    void processVehiclePlaceStateMachine(Company& company);
+}
